Extra init check to make sure value has the correct type
Along with issue #10, a valid Url Picker value must have a type object. If a property has been converted and has a value, the value object may not be in the correct format. To fix, look to see if we have a type object in the value object. If not, initialise as new.